requirements.txt
Flask==3.0.0
spotipy==2.23.0
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
gunicorn==21.2.0